Revert "[Upstream] Upstream is real (#283)"
This reverts commit 22b9372102cf405c906dc24a94259504967f31ee.

+++ Content.Shared/Chat/ChatSelectChannel.cs
@@ -7,7 +7,7 @@
     ///     Maps to <see cref="ChatChannel"/>, giving better names.
     /// </remarks>
     [Flags]
-    public enum ChatSelectChannel : uint
+    public enum ChatSelectChannel : ushort
     {
         None = 0,
